Quantizer Presets update
Now the full functionality of the modal quantizer has returned in the
form of presets. The presets file shared by both (or future) quantizers
in a json file. The scale file is loaded on instantiation. In other
words, it is possible to update and add scales without modifying the
quantizer.