AHD Envelope
AHD Envelope with signal rate controls for attack, hold, delay where 1V
= power of 10.

Formula is 10^(V-1).  All inputs and parameters are signal rate, so you
can trigger as fast as you like.  (max is Nyquist)

The envelope is also velocity sensitive and will scale in proportion to
the pulse.  There's a function (analog) buried in the gen~ patch that
allows the reattacks to get louder if they happen before the envelope
decays; it's a little like the pseudo-dynamics you can create on the
Moog Mini.