Quantized Random Voltages
Clone of the Doepfer a149-1 / quantized random voltages section of the
Buchla 265 Source of Uncertainty.

Has extra controls for time correlation, skew distribution, MSB/LSB,
and extra CV goodness.

Not sure where this goes, though.  Should it be an oscillator?  Doesn't
really seem like it, though "Quantizer" isn't really accurate either.

Maybe a random folder is in order?  (working on the rest of the Source
of Uncertainty)