Added equal subdivision mode to Quantizer
The default mode stays the same ("Near" as in snap to nearest), but the
"equal" mode divides each octave into equal parts.  This is handy when
driving the quantizer with an LFO and will produce smooth scales.

Additionally, a trigger outlet (right) fires off a 5 ms trigger
whenever a new quantized value occurs.  (This is similar to the way the
Topobrillo Quantimator works)

May need some cosmetic tweaks for button layout.