All-in-one Envelope (D/AD/AHD/ASR)
Full-featured all-in-one envelope.  (no ADSR, but we have other ones
for that)  Has retriggering, looping, manual triggering, CV adjustable
attack, hold, decay/release, skew, time scaling, and curve, and outputs
for End of Attack (EOA), End of Hold (EOH), and End of Decay/Release
(EOD/EOR).

It has hidden options for adjusting the min and max attack times as
well as the retrigger duration past the boundary of the bpatcher.